Encode UTF-8 properly, otherwise we can easily trigger unhandled exceptions in analyzer.py and prevent further analysis of the malware.  This has already been observed in practice:
https://malwr.com/analysis/ODE4MTdkMDY5YTEyNDIzNjgzYmVjMDJiODY0ZWY4MzU/#
should have thousands of failed calls to DeleteFileW but are not displayed because cuckoomon is waiting for CallNamedPipe to return, which it never will because analyzer.py will never write/close the pipe handle due to the unhandled exception on UTF-8 decoding.

The entire Cuckoo analyzer code is broken with regards to handling unicode filenames, so a separate patch will be committed to it to fix the additional issues there.

Also fix a typo and disable hooks around reporting exceptions