Fix two bugs in the upstream cuckoomon:
It was missing a hook on NtResumeProcess, so we had unexplained execution of a secondary suspended process in the case of Dyre
It's incorrect to put a hook on ExitProcess as it can result in additional userland code (atexit-type handlers) which then won't be logged because we're executing with ExitProcess in the call trace.  Not only that, but we would shut down the logging system prior to ExitProcess being called, making any further API logs impossible.
Remove the ExitThread hook as well, we'll just log NtTerminateThread