Today I discovered through much wasted time that capstone was unable to correctly determine the size of a trivial instruction used hundreds of times in even ntdll.  Since it somehow managed to get past version 2.0 without being able to reliably perform even the most simple operation expected of a disassembly engine, one which even z0mbie's LDE from 1999 could handle, it's simply not worth using.  I am therefore reverting back to distorm, which is able to handle these very trivial instructions without issue.
For the curious, the instruction was: movzx eax, byte ptr [ebp+8] (0f b6 45 08).  Capstone recognized this instead as a 3 byte instruction, completely breaking the created trampoline for RtlDecompressBuffer on Windows 7.  While it's entirely possible this problem was fixed in version 3.0 of capstone, I can't afford to waste more time finding out what other "surprises" it has in store.