Introduce a number of changes:
Fix crashes in GetDiskFreeSpaceExA and GetDiskFreeSpaceExW, observed during URL analysis.  Caused by assuming the presence of an optional argument.

Improve DLL analysis by fixing our handling of LdrLoadDll.  Now we will treat it as a notail hook, which unfortunately means we'll have no insight into the return value of the call.  It also means initially there will be some more API log clutter that didn't exist previously as only the LdrLoadDll was logged, now we'll see everything called in the process of the LdrLoadDll, including that of the entrypoint code of the DLL.  All the post-DLL loading code and DLL unloading code is now moved to our DllLoadNotification hook created earlier.

Instead of just passing the special flag to enter_hook(), pass the address of the actual hook_t struct.  This will give us the ability to know at enter_hook time what API we're intercepting, allowing for some greater future flexibility.

Extend the unwind info for the notail code as we do actually need to use it, both for the enter_hook and our notail functions.  Expand the size of the pretramp code and add assertions against the size we're crafting.

Fix a number of bugs in the notail assembly, including a missing repne prefix, invalid calls that returned to the address of the call target instead of jumping over it, and don't clobber the rcx value we use for the stack args copying.