Introduce a number of improvements that will allow us to update the cuckoomon binaries without causing any detrimental changes to the API logs:
Implement an addition to the notail hook type: allow the notail function to determine whether we call the original function or a hooked version that will put ourselves on the call stack and force an ignore of anything done within that function.

Noted several instances of tail call optimization preventing us from being located on the call stack, introduced code to prevent that optimization in the affected cases.

Make LdrLoadDll act the same as before except in instances of DLLs that we don't know anything about (yes there's an obvious flaw to it, but we can harden it up in the future if need be).