Ever since we've been using a proper TlsAlloc/GetTlsValue/SetTlsValue, we've reintroduced the same bug that existed with NtAllocateVirtualMemory in the first place that we had the clever hack for.  This one happened to be trickier to reproduce because it required a number of conditions:
1) We had to be executing our NtAllocateVirtualMemory hook
2) It had to be the first hook executed on a new thread
3) We had to be injected into a process that had done enough TLS allocations of its own for us to end up with an index >= 64 (generally this would be observed in Windows 7+ on injected explorer processes depending on how long explorer had been running prior to the VM snapshot)

The reason is due to SetTlsValue (which was used even with our hack) ending up calling RtlAllocateHeap if the TLS index was >= 64.  As this would be called prior to the establishing of a hook_info struct for the current thread, we'd end up in infinite stack recursion which generally were never able to be logged.

This is yet another hack with some failure modes known to me, but should be good enough for our purposes.