Fix an ancient cuckoomon bug that predated any of my changes.  This came to bite us with some recent samples, including a Dridex sample reported by micomico14: https://github.com/brad-accuvant/cuckoomon-modified/issues/28
Though incorrect code in general, it wasn't visible until just now.  The problem lied in blindly using the result of GetModuleHandle on the library name for any hook about to be applied.  The problem is that when hooks are attempted to be applied against DLLs that have not yet loaded, GetModuleHandle will return NULL.  Like other similar APIs, NULL is a valid module handle for GetProcAddress and will cause the main executable to be selected by BasepMapModuleHandle.  In most cases, main executables don't contain any exports, or if any did, it was unlikely that they duplicated standard API names for their exported functions.  So it wasn't noticed until just recently when Dridex started supplying binaries with anomalous export directories.  In this case, GetProcAddress would attempt to walk the bogus export table and end up crashing on following garbage pointers, preventing cuckoomon from being able to load in the process (but permitting malicious activity to continue unlogged).

It's not clear whether this was simply an ancient bug whose time had come to be exposed or if this was a targeted anti-Cuckoo attempt.