More correctly deal with the issue mentioned at: http://joe4security.blogspot.com/2016/07/rise-of-vbs-evading-sandboxes.html
The analysis is a bit wrong (it'd only delay execution for 270 seconds,
the Wscript.sleep time is irrelevant and not measured in seconds anyway)
and the use of timer() isn't really interesting (it's just a GetLocalTime
wrapper), but rather the Wscript.sleep implementation which performs an
effective sleep through win32k.sys instead of the normal
NtDelayExecution/etc sleeps.  Copy over some of our existing
NtDelayExecution handling to deal with it.